+++ tests/unit/Fixture/Definition/ClassDefinitionWithPropertyDependencies.php
@@ -1,10 +1,10 @@
 <?php
-        $entry = new \A();
+        $entry = new \X\A();
         $this->setProperties(
             $entry,
             [
-                'b' => $this->singletonEntries['B'] ?? $this->B(),
-                'c' => $this->singletonEntries['C'] ?? $this->C(),
+                'b' => $this->singletonEntries['X\B'] ?? $this->X__B(),
+                'c' => $this->singletonEntries['X\C'] ?? $this->X__C(),
             ]
         );