+++ tests/unit/Fixture/Definition/ClassDefinitionWithRequiredConstructorDependencies.php
@@ -1,5 +1,5 @@
 <?php
-        return new \A(
-            $this->singletonEntries['B'] ?? $this->B(),
-            $this->singletonEntries['C'] ?? $this->C()
+        return new \X\A(
+            $this->singletonEntries['X\B'] ?? $this->X__B(),
+            $this->singletonEntries['X\C'] ?? $this->X__C()
         );