+++ tests/unit/Fixture/Definition/ReferenceDefinitionPrototype.php
@@ -1,2 +1,2 @@
 <?php
-        return $this->singletonEntries['B'] ?? $this->B();
+        return $this->singletonEntries['X\B'] ?? $this->X__B();